+++ bukkit/src/main/java/de/maxhenkel/voicechat/voice/common/Utils.java
@@ -1,5 +1,7 @@
 package de.maxhenkel.voicechat.voice.common;
 
+import de.maxhenkel.voicechat.Voicechat;
+
 public class Utils {
 
     public static short[] bytesToShorts(byte[] bytes) {
@@ -67,6 +69,15 @@ public class Utils {
             }
         }
         return floats;
+    }
+
+    /**
+     * Gets the default voice chat distance
+     *
+     * @return the default voice chat distance
+     */
+    public static float getDefaultDistance() {
+        return Voicechat.SERVER_CONFIG.voiceChatDistance.get().floatValue();
     }
 
 }